+++ tests/table/test_tx.py
@@ -89,13 +89,32 @@ def test_tx_snapshot_ro(driver_sync, database):
     assert data[0].rows == [{"value": 2}]
 
 
-def test_split_transactions_deny_split(driver_sync, table_name):
+def test_split_transactions_deny_split_explicit_commit(driver_sync, table_name):
     with ydb.SessionPool(driver_sync, 1) as pool:
 
         def check_transaction(s: ydb.table.Session):
             with s.transaction(allow_split_transactions=False) as tx:
                 tx.execute("INSERT INTO %s (id) VALUES (1)" % table_name)
                 tx.commit()
+
+                with pytest.raises(RuntimeError):
+                    tx.execute("INSERT INTO %s (id) VALUES (2)" % table_name)
+
+                tx.commit()
+
+            with s.transaction() as tx:
+                rs = tx.execute("SELECT COUNT(*) as cnt FROM %s" % table_name)
+                assert rs[0].rows[0].cnt == 1
+
+        pool.retry_operation_sync(check_transaction)
+
+
+def test_split_transactions_deny_split_flag_commit(driver_sync, table_name):
+    with ydb.SessionPool(driver_sync, 1) as pool:
+
+        def check_transaction(s: ydb.table.Session):
+            with s.transaction(allow_split_transactions=False) as tx:
+                tx.execute("INSERT INTO %s (id) VALUES (1)" % table_name, commit_tx=True)
 
                 with pytest.raises(RuntimeError):
                     tx.execute("INSERT INTO %s (id) VALUES (2)" % table_name)